requirements.txt
pandas
openpyxl
sqlalchemy
fastapi
uvicorn[standard]
python-multipart
python-dotenv
opencv-python
pyzbar
pillow
pytesseract
